validation before servlets is going to be done. Think about checks in servlets, how can i get to them except the filter way

+++ src/main/java/ru/senina/itmo/web/web_lab_2/entities/Coordinates.java
@@ -14,7 +14,7 @@ public class Coordinates implements Serializable {
     @Getter
     private double r;
 
-    public Coordinates(int x, double y, double r) {
+    public Coordinates(double x, double y, double r) {
         setX(x);
         setY(y);
         setR(r);
